format and remove unused code

+++ src/main/java/com/quanqinle/epub/MakeEpubFromTemplate.java
@@ -27,81 +27,103 @@ public class MakeEpubFromTemplate {
 
     public static final Logger logger = LoggerFactory.getLogger(MakeEpubFromTemplate.class);
 
+    BookInfo book;
     /**
      * original plain text file, the full path can be read directly
      */
     Path srcFilePath;
-    BookInfo book;
     /**
      * the full path of epub template, copy from it, do not modify the original files
      */
-    URL templateSrcUrl;
+    private Path templateSrcPath;
     /**
      * copy epub template to this folder, then modify it
      */
-    Path templateDstPath;
+    private Path templateDstPath;
 
     /**
      * `<navPoint></navPoint>` list in `toc.ncx`
      */
-    String navPointList = "";
+    private String navPointList = "";
     /**
      * `<item></item>` list in `content.opf`
      */
-    String itemList = "";
+    private String itemList = "";
     /**
      * `<itemref></itemref>` list in `content.opf`
      */
-    String itemrefList = "";
+    private String itemrefList = "";
     /**
      * `<reference></reference>` list in `content.opf`
      */
-    String referenceList = "";
+    private String referenceList = "";
     /**
      * `<li></li>` list in `toc.xhtml`
      */
-    String tocItemList = "";
+    private String tocItemList = "";
 
     public static void main(String[] args) throws IOException, URISyntaxException {
-        Path srcFilePath = Paths.get("D:", "book-library", "demo.txt");
+        Path srcFilePath = Paths.get("D:", "book-library", "DEMO.txt");
 
         BookInfo book = new BookInfo();
         book.setOutputDir(Paths.get("D:", "epub"));
         book.setCoverJpgFullPath(Paths.get("D:", "book.jpg"));
-        book.setUUID(UUID.randomUUID().toString());
+        book.setUuid(UUID.randomUUID().toString());
         book.setLanguage("zh");
         book.setBookTitle("红楼梦");
         book.setAuthor("曹雪芹");
         book.setCreateDate("2021-03-06");
 
         /*
-         the following is a complete demo, making srcFilePath to a .epub book
+         * demo: make srcFilePath to a .epub book
          */
         MakeEpubFromTemplate makeEpub = new MakeEpubFromTemplate(srcFilePath, book);
-        makeEpub.copyTemplate();
-        makeEpub.genBodyHtmls();
-        makeEpub.setBookCover();
-        makeEpub.setBookTocHtml();
-        makeEpub.modifyTocNcx();
-        makeEpub.modifyContentOpf();
-
-        makeEpub.zipEpub();
+        makeEpub.make();
     }
 
     /**
      *
-     * @param srcFilePath - original plain text file
-     * @param book -
+     * @param srcFilePath original plain text file
+     * @param book
      */
-    public MakeEpubFromTemplate(Path srcFilePath, BookInfo book) {
+    public MakeEpubFromTemplate(Path srcFilePath, BookInfo book)  {
         this.srcFilePath = srcFilePath;
         this.book = book;
-        this.templateDstPath = book.getOutputDir().resolve(Constant.templateName);
-        this.templateSrcUrl = this.getClass().getClassLoader().getResource(Constant.templateName);
+        this.templateDstPath = book.getOutputDir().resolve(Constant.TEMPLATE_NAME);
+
+        try {
+            URL templateSrcUrl = getClass().getClassLoader().getResource(Constant.TEMPLATE_NAME);
+            assert templateSrcUrl != null;
+            this.templateSrcPath = Paths.get(templateSrcUrl.toURI());
+        } catch (URISyntaxException e) {
+            logger.error("Fail to find the epub source template: {}", templateSrcPath);
+            System.exit(0);
+        }
+
+    }
+
+    /**
+     * start to make the .epub file.
+     *
+     * The method is an all in one method, it includes read-parse-rewrite, etc.,
+     * so you can use it to make a .epub book just after the construction method.
+     *
+     * @throws IOException
+     */
+    public void make() throws IOException {
+        copyTemplate();
+        genBodyHtmls();
+        setBookCover();
+        setBookTocHtml();
+        modifyTocNcx();
+        modifyContentOpf();
+
+        zipEpub();
     }
 
     /**
      * zip folder to epub.
+     *
      * Use {@link com.quanqinle.epub.util.EpubUtils#zipEpub(Path epubSrcFolderPath, Path epubFilePath)}
      * Folder is from the copy of original template.
      * epub is from book.outputdir+book.title+.epub.
@@ -113,22 +135,22 @@ public class MakeEpubFromTemplate {
         if (null == bookName || bookName.isBlank()) {
             bookName = "ebook";
         }
-        EpubUtils.zipEpub(this.getTemplateDstPath(), book.getOutputDir().resolve(bookName + ".epub"));
+        EpubUtils.zipEpub(this.templateDstPath, book.getOutputDir().resolve(bookName + ".epub"));
     }
 
 
     /**
-     * 拷贝epub模板
+     * copy epub source template to the output directory.
+     * This method will delete (if exist) and create the output directory.
      */
-    public void copyTemplate() throws IOException, URISyntaxException {
+    public void copyTemplate() throws IOException {
 
-        Path src = Paths.get(templateSrcUrl.toURI());
-//        Files.walk(src).forEach(System.out::println);
-
+        Path src = templateSrcPath;
         Path dst = templateDstPath;
+//        Files.walk(src).forEach(System.out::println); // debug
 
-        logger.info("src={}", src);
-        logger.info("dst={}", dst);
+        logger.info("source template = {}", src);
+        logger.info("output folder = {}", dst);
 
         if (Files.exists(dst)) {
             Files.walk(dst)
@@ -151,6 +173,11 @@ public class MakeEpubFromTemplate {
                 );
     }
 
+    /**
+     * Modify cover.xhtml
+     *
+     * @throws IOException
+     */
     public void setBookCover() throws IOException {
         Path coverHtml = templateDstPath.resolve("OEBPS/Text/cover.xhtml");
         String content =Files.readString(coverHtml);
@@ -159,7 +186,7 @@ public class MakeEpubFromTemplate {
 
         Path src = book.getCoverJpgFullPath();
         Path dst = templateDstPath.resolve("OEBPS/Images/cover.jpg");
-        if (src != null) {
+        if (src != null && Files.exists(src)) {
             Files.copy(src, dst, StandardCopyOption.REPLACE_EXISTING);
         } else {
             // TODO
@@ -167,7 +194,7 @@ public class MakeEpubFromTemplate {
     }
 
     /**
-     * 设置 toc.xhtml
+     * Modify toc.xhtml
      *
      * @throws IOException -
      */
@@ -180,7 +207,7 @@ public class MakeEpubFromTemplate {
     }
 
     /**
-     * generate HTML files of the book,
+     * Generate HTML files of the book,
      * and init some parameters which could be used when setting toc.xhtml, toc.ncx, content.opf, etc.
      *
      * @author quanqinle
@@ -193,13 +220,13 @@ public class MakeEpubFromTemplate {
     }
 
     /**
-     * toc.ncx
+     * Modify toc.ncx
      */
     public void modifyTocNcx() throws IOException {
         Path tocNcxPath = templateDstPath.resolve("OEBPS/toc.ncx");
 
         String content = Files.readString(tocNcxPath);
-        content = content.replace("[ISBN]", book.getISBN())
+        content = content.replace("[ISBN]", book.getIsbn())
                 .replace("[BOOK'S TITLE]", book.getBookTitle())
                 .replace("[navPoint LIST]", this.navPointList);
 
@@ -207,14 +234,14 @@ public class MakeEpubFromTemplate {
     }
 
     /**
-     * content.opf
+     * Modify content.opf
      */
     public void modifyContentOpf() throws IOException {
         Path contentOpfPath = templateDstPath.resolve("OEBPS/content.opf");
 
         String content =Files.readString(contentOpfPath);
-        content = content.replace("[UUID]", book.getUUID())
-                .replace("[ISBN]", book.getISBN())
+        content = content.replace("[UUID]", book.getUuid())
+                .replace("[ISBN]", book.getIsbn())
                 .replace("[BOOK'S TITLE]", book.getBookTitle())
                 .replace("[NAME LASTNAME]", book.getAuthor())
                 .replace("[LASTNAME, NAME]", book.getAuthor())
@@ -233,24 +260,24 @@ public class MakeEpubFromTemplate {
     /**
      * organize contents for toc.ncx and content.opf.
      *
-     * This method has been used in `genBodyHtmls()`
+     * This method has been used in {@link #genBodyHtmls()}
      */
     private void makeContentForTocNcxAndContentOpf() {
         int index = 1;
 
-        navPointList = navPointList.concat(String.format(Constant.navPointFormat, index, index, "封面", "cover.xhtml"));
-        referenceList = referenceList.concat(String.format(Constant.referenceFormat, "cover", "cover.xhtml", "封面"));
-        tocItemList = tocItemList.concat(String.format(Constant.tocItemFormat, "cover.xhtml", "封面"));
+        navPointList = navPointList.concat(String.format(Constant.FORMAT_NAV_POINT, index, index, "封面", "cover.xhtml"));
+        referenceList = referenceList.concat(String.format(Constant.FORMAT_REFERENCE, "cover", "cover.xhtml", "封面"));
+        tocItemList = tocItemList.concat(String.format(Constant.FORMAT_TOC_ITEM, "cover.xhtml", "封面"));
         index++;
 
-        navPointList = navPointList.concat(String.format(Constant.navPointFormat, index, index, "目录", "toc.xhtml"));
-        referenceList = referenceList.concat(String.format(Constant.referenceFormat, "toc", "toc.xhtml", "目录"));
-        tocItemList = tocItemList.concat(String.format(Constant.tocItemFormat, "toc.xhtml", "目录"));
+        navPointList = navPointList.concat(String.format(Constant.FORMAT_NAV_POINT, index, index, "目录", "toc.xhtml"));
+        referenceList = referenceList.concat(String.format(Constant.FORMAT_REFERENCE, "toc", "toc.xhtml", "目录"));
+        tocItemList = tocItemList.concat(String.format(Constant.FORMAT_TOC_ITEM, "toc.xhtml", "目录"));
         index++;
 
         int i = 0;
         for (String chapterTitle: book.getHtmlFileMap().keySet()) {
-            if (0 == i && !Constant.frontMatterTitle.equals(chapterTitle)) {
+            if (0 == i && !Constant.FRONT_MATTER_TITLE.equals(chapterTitle)) {
                 // the chapter just before the 1st chapter
                 i++;
                 continue;
@@ -259,11 +286,11 @@ public class MakeEpubFromTemplate {
             FileInfo file = book.getHtmlFileMap().get(chapterTitle);
             String fileName = file.getName();
 
-            navPointList = navPointList.concat(String.format(Constant.navPointFormat, index, index, chapterTitle, fileName));
-            itemList = itemList.concat(String.format(Constant.itemFormat, fileName, i));
-            itemrefList = itemrefList.concat(String.format(Constant.itemrefFormat, i));
-            referenceList = referenceList.concat(String.format(Constant.referenceFormat, "text", fileName, chapterTitle));
-            tocItemList = tocItemList.concat(String.format(Constant.tocItemFormat, fileName, chapterTitle));
+            navPointList = navPointList.concat(String.format(Constant.FORMAT_NAV_POINT, index, index, chapterTitle, fileName));
+            itemList = itemList.concat(String.format(Constant.FORMAT_ITEM, fileName, i));
+            itemrefList = itemrefList.concat(String.format(Constant.FORMAT_ITEMREF, i));
+            referenceList = referenceList.concat(String.format(Constant.FORMAT_REFERENCE, "text", fileName, chapterTitle));
+            tocItemList = tocItemList.concat(String.format(Constant.FORMAT_TOC_ITEM, fileName, chapterTitle));
 
             index++;
             i++;
@@ -271,11 +298,4 @@ public class MakeEpubFromTemplate {
 
     }
 
-    public Path getTemplateDstPath() {
-        return templateDstPath;
-    }
-
-    public void setTemplateDstPath(Path templateDstPath) {
-        this.templateDstPath = templateDstPath;
-    }
 }
